authorizationToken to enable initiation by PISP
The european regulation authorize regulated entities (Payment Initiation Service Provider) to initiate the payment under  the mandate of the owner of the account. To enable this regulatory provision, the authorization token is set by the Payer's bank during the check-out process, put in the response and will be used shorly after in a bilateral mode between the PISP and the payer's bank (namend ASPSP in the european regulation, for Account Service Payment Service Provider).